requirements.txt
python-multipart
pydantic
gunicorn
uvicorn
fastapi
python-dotenv
pymongo
tenacity
groq
requests
together 
celery
redis